+++ packages/idyll-cli/src/pipeline/index.js
@@ -72,7 +72,7 @@ const build = (opts, paths, resolvers) => {
           }
         };
         if (!opts.ssr) {
-          output.html = getBaseHTML(ast, template);
+          output.html = getBaseHTML(ast, template, opts);
         } else {
           output.html = getHTML(
             paths,
@@ -80,11 +80,7 @@ const build = (opts, paths, resolvers) => {
             output.components,
             output.data,
             template,
-            {
-              ssr: opts.ssr,
-              theme: opts.theme,
-              layout: opts.layout
-            })
+            opts)
         }
       })
     })