+++ packages/idyll-cli/test/multiple-component-dirs/test.js
@@ -73,6 +73,10 @@ beforeAll(done => {
 
 test('options work as expected', () => {
   expect(idyll.getOptions()).toEqual({
+    alias: {
+      'PackageJsonComponentTest': 'CustomComponent'
+    },
+    context: undefined,
     layout: 'centered',
     theme: join(PROJECT_DIR, 'custom-theme.css'),
     minify: false,
@@ -94,14 +98,7 @@ test('options work as expected', () => {
     compiler: {
       spellcheck: false
     },
-    inputString: fs.readFileSync(join(PROJECT_DIR, 'index.idl'), 'utf-8'),
-    inputConfig: {
-      compiler: {},
-      components: {
-        'package-json-component-test': 'components-1/custom-component.js'
-      },
-      transform: []
-    }
+    inputString: fs.readFileSync(join(PROJECT_DIR, 'index.idl'), 'utf-8')
   })
 })