+++ packages/idyll-compiler/src/index.js
@@ -3,14 +3,12 @@ const parse = require('./parser');
 const Lexer = require('./lexer');
 const Processor = require('./processors');
 const { cleanNewlines } = require('./processors/pre');
-const { hoistVariables, flattenChildren, cleanResults, makeFullWidth, wrapText } = require('./processors/post');
+const { hoistVariables, flattenChildren, cleanResults, makeFullWidth, wrapText, autoLinkify } = require('./processors/post');
 const matter = require('gray-matter');
 
-module.exports = function(input, options, callback) {
-  input = Processor(input).pipe(cleanNewlines).end();
-
+module.exports = function(input, options, callback) { 
+  input = Processor(input).pipe(cleanNewlines).end();  
   const { content, data } = matter(input.trim());
-
   options = Object.assign({}, { spellcheck: false, smartquotes: true, async: true }, options || {});
   const lex = Lexer();
   let lexResults = '', output = [];
@@ -26,13 +24,14 @@ module.exports = function(input, options, callback) {
     console.warn(`\nError parsing Idyll markup:\n${err.message}`);
     return new Promise((resolve, reject) => reject(err));
   }
-
-  let astTransform = Processor(output, options)
+ 
+  let astTransform = Processor(output, options) 
     .pipe(hoistVariables)
     .pipe(flattenChildren)
     .pipe(makeFullWidth)
     .pipe(wrapText)
     .pipe(cleanResults)
+    .pipe(autoLinkify)
     .end();
 
   if (options.postProcessors) {