+++ packages/idyll-docs/gallery/contents.js
@@ -4,6 +4,13 @@ const exampleGroups = [
     title: 'Explorable Explanations',
     examples: [
       {
+        label: "The Beginner's Guide to Dimensionality Reduction",
+        subtitle: "Explore the methods that data scientists use to visualize high-dimensional data.",
+        href: 'https://idyll.pub/post/dimensionality-reduction-293e465c2a3443e8941b016d/',
+        image: 'https://idyll.pub/post/dimensionality-reduction-293e465c2a3443e8941b016d/static/images/share.png',
+        sourceUrl: "https://github.com/mathisonian/dimensionality-reduction"
+      },
+      {
         label: 'The Math of Card Shuffling',
         subtitle: "Riffling from factory order to complete randomness.",
         href: 'https://fredhohman.com/card-shuffling/',
@@ -84,6 +91,13 @@ const exampleGroups = [
   {
     title : 'Articles and Blog Posts',
     examples: [
+      {
+        label: 'A Comic Introduction to Idyll',
+        subtitle: 'See how Idyll can be used with CSS grid to create a comic-style layout.',
+        href: 'https://mathisonian.github.io/idyll-comic/',
+        image: 'https://mathisonian.github.io/idyll-comic/static/share.png',
+        sourceUrl: "https://github.com/mathisonian/idyll-comic/"
+      },
       {
         label: 'Autumn Colormaps',
         subtitle: 'A look at trees, colormaps, and chroma.js',