+++ packages/idyll-docs/idyll-components/contents.js
@@ -54,6 +54,9 @@ const components = [
             "idyllProps": [
               {
                 "currentStep": "the index of the currently selected step."
+              },
+              {
+                "currentState": "the state value associated with the currently selected step. Note you must set the state property on the step components for this value to update."
               }
             ]
           }
@@ -217,6 +220,23 @@ const components = [
               },
               {
                 "type": "The type of the chart to display, can be `line`, `scatter`, `bar`, `pie`, or `time`. The time type is a line chart that expects the `x` values in the data to be in the temporal domain."
+              },
+              {
+                "domain": "The chart extent along the x dimension."
+              },
+              {
+                "range": "The chart extent along the y dimension."
+              }
+            ]
+          }
+        },
+        {
+          "Conditional": {
+            "thumbnail": "conditional.png",
+            "description": "This component will conditionally display its children.",
+            "idyllProps": [
+              {
+                "if": "An expression; if this evaluates to true, the children will be rendered, otherwise nothing will be drawn to the screen"
               }
             ]
           }