+++ packages/idyll-docs/pages/docs/configuration-and-styles.js
@@ -1,34 +1,16 @@
 import Link from 'next/link'
 import markdown from 'markdown-in-js'
 import Layout from '../../components/layout'
-
+import Highlight from 'react-highlight'
 
 const Content = () => markdown`
-# Configuration and Styles
-
-## Command line options
-
-The \`idyll\` command line tool accepts the following options
-
-* \`--components\` the path to your custom components. By default this points to \`components/\`. This can be a string (a single path), or an array.
-* \`--defaultComponents\` the path to Idyll's built-in components. By default this isn't set and it uses the latest package from npm.
-* \`--css\` the path to your CSS file. You can use this to override Idyll's default styles, e.g. \`$ idyll index.idl --css my-custom-styles.css\`.
-* \`--datasets\` the path to the folder containing your datasets. By default this points to \`data/\`.
-* \`--inputString\` a string of Idyll markup. Use this flag to compile a string instead of a file.
-* \`--layout\` the name of the layout to use. By default this is \`blog\`. More on layouts below.
-* \`--no-minify\` turn code minification off when building for production.
-* \`--no-ssr\` turn server-side rendering off when building for production.
-* \`--output\` the folder in which to place build output.
-* \`--theme\` the name of the theme to use. By default this is \`idyll\`. More on themes below.
-* \`--template\` a custom HTML template to use. The default template is can be found [here](https://github.com/idyll-lang/idyll/blob/master/packages/idyll-cli/src/client/_index.html).
-* \`--watch\` the watch flag tells Idyll if it should watch files for changes, or just run once and exit.
-
-If you are using Idyll via the project generator, open \`package.json\` to change these options.
 
 ## Themes and page layout
 
 Idyll exposes two options to help you style your project, \`layout\` and \`theme\`. \`layout\` deals with CSS styles related to how your content is
 layed out on the page: width, columns, etc. The \`theme\` option allows you to choose different stylesheets to change the style of the content itself (text color, font, and so on).
+
+If more customization is required, you can provide a custom HTML template, and Idyll will generate articles within that custom template (see the "template" option above).
 
 ### Layout
 
@@ -94,10 +76,52 @@ idyll
   .on('update', () => { console.log('Finished.') });
 `
 
-
-
 export default ({ url }) => (
-  <Layout url={ url }>
+  <Layout url={ url } title={'Idyll Documentation | Configuration and Styles'}>
+    <link rel="stylesheet" href="../../static/styles/tomorrow-night-eighties.css" />
+
+    <h1>Configuration and Styles</h1>
+
+    <p>Idyll is typically configured via options set in package.json, but all these same options can be configured via command line flags or when using Idyll as an API.</p>
+
+    <h2>Options</h2>
+
+
+    <ul>
+      <li><b>layout</b> - the name of the layout to use. By default this is "blog". More on layouts below.</li>
+      <li><b>theme</b> - the name of the theme to use. By default this is "github". More on themes below.</li>
+      <li><b>googleFonts</b> - a list of Google font names to include via CSS. This can be an array of strings or a single string.</li>
+      <li><b>favicon</b> - an ico file to use as the favicon, should be inside of the static folder, e.g. "static/favicon.ico".</li>
+      <li><b>alias</b> - Customize component resolution, for example <code>{`{ "VL": "IdyllVegaLite" }`}</code> would let you use tags like <code>{`[VL /]`}</code> in your markup, instead of <code>{`[IdyllVegaLite /]`}</code>.</li>
+      <li><b>template</b> - a custom HTML template to use. The default template can be found <a href="https://github.com/idyll-lang/idyll/blob/master/packages/idyll-cli/src/client/_index.html">here</a></li>
+      <li><b>components</b> - the path to your custom components. By default this points to "components". This can be a string (a single path), or an array.</li>
+      <li><b>css</b> - the path to a CSS file. You can use this to override Idyll's default styles.</li>
+      <li><b>datasets</b> - the path to the folder containing your datasets. By default this points to "data/".</li>
+      <li><b>no-minify</b> - turn code minification off when building for production.</li>
+      <li><b>no-ssr</b> - turn server-side rendering off when building for production.</li>
+      <li><b>output</b> - the folder in which to place build output..</li>
+      <li><b>inputString</b> - string of Idyll markup. Use this flag to compile a string instead of a file.</li>
+    </ul>
+
+    For example, package.json might contain the following:
+
+    <Highlight className='json'>
+{`"idyll": {
+  "layout": "blog",
+  "theme": "my-custom-theme.css",
+  "components": ["./components/", "../some-other-components/"],
+  "googleFonts": ["Hanalei Fill"],
+  "favicon": "static/favicon.ico",
+  "alias": {
+    "VL": "IdyllVegaLite"
+  }
+}`}
+    </Highlight>
+
+    <p>
+      <em>Note, when modifying these options you may need to restart the idyll server for the change to take effect.</em>
+    </p>
+
     <Content />
 
     <section>
@@ -107,9 +131,9 @@ export default ({ url }) => (
         For example, you could make a static blog engine that uses Idyll to compile the blog posts.
       </p>
       <h3>Example</h3>
-      <pre>
-        <code className="lang-js">{ ExampleCodeA }</code>
-      </pre>
+      <Highlight className='javascript'>
+        { ExampleCodeA }
+      </Highlight>
       <p>
         If you pass <code>live: true</code> to Idyll, it will continue to watch the input files for changes,
         and will emit the update event each time that the output is rebuilt.
@@ -117,9 +141,9 @@ export default ({ url }) => (
       <p>
         You can also compile an input string directly instead of a file:
       </p>
-      <pre>
-        <code className="lang-js">{ ExampleCodeB }</code>
-      </pre>
+      <Highlight className='javascript'>
+        { ExampleCodeB }
+      </Highlight>
     </section>
 
     <p>