+++ packages/idyll-docs/pages/docs/publishing/embedding.js
@@ -1,12 +1,13 @@
 import Link from 'next/link'
 import markdown from 'markdown-in-js'
 import Layout from '../../../components/layout'
-
-
+import Highlight from 'react-highlight'
 
 export default ({ url }) => (
-  <Layout url={ url }>
+  <Layout url={url} title={'Idyll Documentation | Embedding Idyll on a webpage'}>
+    <link rel="stylesheet" href="../../../static/styles/tomorrow-night-eighties.css" />
     <div>
+
       <h1>Embedding Idyll in an existing web page</h1>
 
       <p>The Idyll runtime is available as a React
@@ -15,14 +16,14 @@ export default ({ url }) => (
 
       <p>To do this, you must first install the dependencies:</p>
 
-      <pre><code class="sh language-sh">$ npm i --save idyll-document idyll-components
+      <pre><code>$ npm i --save idyll-document idyll-components
       </code></pre>
 
       <p>then, add it to your page. If you are already using React, you
       can include this as a standard component:</p>
 
-      <pre><code class="javascript language-javascript">
-{`import IdyllDocument from 'idyll-document';
+      <Highlight className='javascript'>
+        {`import IdyllDocument from 'idyll-document';
 import * as components from 'idyll-components';
 
 // An example functional component
@@ -35,17 +36,17 @@ import * as components from 'idyll-components';
   )
 }
 `}
-      </code></pre>
+      </Highlight>
 
       <p>If not, you'll also need to install <code>react</code> and <code>react-dom</code>:</p>
 
-      <pre><code class="sh language-sh">$ npm i --save react react-dom
-      </code></pre>
+      <Highlight className='javascript'>$ npm i --save react react-dom
+      </Highlight>
 
       <p>and can embed it like this:</p>
 
-      <pre><code class="javascript language-javascript">
-{`import React from 'react';
+      <Highlight className='javascript'>
+        {`import React from 'react';
 import ReactDOM from 'react-dom';
 import IdyllDocument from 'idyll-document';
 import * as components from 'idyll-components';
@@ -60,7 +61,8 @@ ReactDOM.render(
   containerElement
 )
 `}
-      </code></pre>
+      </Highlight>
     </div>
   </Layout>
 )
+